UHF-8525: Fixed kernel test
