+++ tests/src/Unit/EventSubscriber/CacheInvalidatorSubscriberTest.php
@@ -7,6 +7,7 @@ namespace Drupal\Tests\helfi_api_base\Unit\EventSubscriber;
 use Drupal\Core\Cache\CacheTagsInvalidatorInterface;
 use Drupal\helfi_api_base\Azure\PubSub\PubSubMessage;
 use Drupal\helfi_api_base\EventSubscriber\CacheTagInvalidatorSubscriber;
+use Drupal\Tests\helfi_api_base\Traits\CacheTagInvalidatorTrait;
 use Drupal\Tests\UnitTestCase;
 use Prophecy\PhpUnit\ProphecyTrait;
 
@@ -17,6 +18,7 @@ use Prophecy\PhpUnit\ProphecyTrait;
 class CacheInvalidatorSubscriberTest extends UnitTestCase {
 
   use ProphecyTrait;
+  use CacheTagInvalidatorTrait;
 
   /**
    * @covers ::getSubscribedEvents
@@ -38,54 +40,6 @@ class CacheInvalidatorSubscriberTest extends UnitTestCase {
     $invalidator->invalidateTags(['node:123'])->shouldBeCalled();
     $sut = new CacheTagInvalidatorSubscriber($invalidator->reveal());
     $sut->onReceive(new PubSubMessage(['data' => ['tags' => ['node:123']]]));
-  }
-
-  /**
-   * Mock core's CacheTagsInvalidator class.
-   *
-   * The CacheTagsInvalidatorInterface does not define 'resetChecksums()'
-   * method and since the default CacheTagsInvalidator class is marked
-   * as final we cannot mock it.
-   *
-   * @return \Drupal\Core\Cache\CacheTagsInvalidatorInterface
-   *   The cache invalidator class.
-   */
-  private function mockCacheInvalidator() : CacheTagsInvalidatorInterface {
-    return new class () implements CacheTagsInvalidatorInterface {
-
-      /**
-       * A list of invalidated tags.
-       *
-       * @var array
-       */
-      public array $tags = [];
-
-      /**
-       * The number of times resetChecksums has been called.
-       *
-       * @var int
-       */
-      public int $checkSumResets = 0;
-
-      /**
-       * {@inheritdoc}
-       */
-      public function invalidateTags(array $tags) : void {
-        foreach ($tags as $tag) {
-          $this->tags[$tag] = $tag;
-        }
-      }
-
-      /**
-       * Add missing resetChecksums() method.
-       *
-       * @see \Drupal\Core\Cache\CacheTagsInvalidator::resetChecksums()
-       */
-      public function resetChecksums() : void {
-        $this->checkSumResets++;
-      }
-
-    };
   }
 
   /**
